+++ src/main/java/org/trostheide/generativeArt/core/ParameterType.java
@@ -1,9 +1,0 @@
-package org.trostheide.generativeArt.core;
-
-public enum ParameterType {
-    INTEGER,
-    DOUBLE,
-    BOOLEAN,
-    STRING,
-    COLOR // For future use
-}